update generateApi to work with post methods
