+++ packages/client/src/client-v2/api/getAssets.ts
@@ -3,12 +3,19 @@ import { ClientState } from "../state";
 import { api } from "./generateApi";
 
 export const { request: getAssets, requestWithCancel: getAssetsWithCancel } =
-  api("getAssets", "/v2/fungible/assets", (res, options) => {
-    if (
-      options.includeEvmAssets &&
-      options.includeSvmAssets &&
-      options.includeCw20Assets
-    ) {
-      ClientState.skipAssets = res.chainToAssetsMap as Record<string, Asset[]>;
-    }
+  api({
+    methodName: "getAssets",
+    path: "/v2/fungible/assets",
+    onSuccess: (res, options) => {
+      if (
+        options.includeEvmAssets &&
+        options.includeSvmAssets &&
+        options.includeCw20Assets
+      ) {
+        ClientState.skipAssets = res.chainToAssetsMap as Record<
+          string,
+          Asset[]
+        >;
+      }
+    },
   });
