+++ packages/client/src/client-v2/api/getAssetsBetweenChains.ts
@@ -3,7 +3,7 @@ import { api } from "./generateApi";
 export const {
   request: getAssetsBetweenChains,
   requestWithCancel: getAssetsBetweenChainsWithCancel,
-} = api(
-  "fungibleAssetsBetweenChainsCreate",
-  "/v2/fungible/assets_between_chains",
-);
+} = api({
+  methodName: "fungibleAssetsBetweenChainsCreate",
+  path: "/v2/fungible/assets_between_chains",
+});
