+++ packages/client/src/client-v2/api/getAssetsFromSource.ts
@@ -3,4 +3,7 @@ import { api } from "./generateApi";
 export const {
   request: getAssetsFromSource,
   requestWithCancel: getAssetsFromSourceWithCancel,
-} = api("getAssetsFromSource", "/v2/fungible/assets_from_source");
+} = api({
+  methodName: "getAssetsFromSource",
+  path: "/v2/fungible/assets_from_source",
+});
