+++ packages/client/src/client-v2/api/getBalances.ts
@@ -3,4 +3,7 @@ import { api } from "./generateApi";
 export const {
   request: getBalances,
   requestWithCancel: getBalancesWithCancel,
-} = api("getBalances", "/v2/info/balances");
+} = api({
+  methodName: "getBalances",
+  path: "/v2/info/balances",
+});
