+++ packages/client/src/client-v2/api/getBridges.ts
@@ -1,4 +1,7 @@
 import { api } from "./generateApi";
 
 export const { request: getBridges, requestWithCancel: getBridgesWithCancel } =
-  api("getBridges", "/v2/info/bridges");
+  api({
+    methodName: "getBridges",
+    path: "/v2/info/bridges",
+  });
