+++ packages/client/src/client-v2/api/getChains.ts
@@ -2,8 +2,12 @@ import { ClientState } from "../state";
 import { api } from "./generateApi";
 
 export const { request: getChains, requestWithCancel: getChainsWithCancel } =
-  api("getChains", "/v2/info/chains", (res, options) => {
-    if (options?.includeEvm && options?.includeSvm) {
-      ClientState.skipChains = res.chains;
-    }
+  api({
+    methodName: "getChains",
+    path: "/v2/info/chains",
+    onSuccess: (res, options) => {
+      if (options?.includeEvm && options?.includeSvm) {
+        ClientState.skipChains = res.chains;
+      }
+    },
   });
