+++ packages/client/src/client-v2/api/postMessages.ts
@@ -3,4 +3,8 @@ import { api } from "./generateApi";
 export const {
   request: postMessages,
   requestWithCancel: postMessagesWithCancel,
-} = api("getMsgsV2", "/v2/fungible/msgs");
+} = api({
+  methodName: "getMsgsV2",
+  method: "post",
+  path: "/v2/fungible/msgs",
+});
