+++ packages/client/src/client-v2/api/postRoute.ts
@@ -1,4 +1,8 @@
 import { api } from "./generateApi";
 
 export const { request: postRoute, requestWithCancel: postRouteWithCancel } =
-  api("getRouteV2", "/v2/fungible/route");
+  api({
+    methodName: "getRouteV2",
+    method: "post",
+    path: "/v2/fungible/route",
+  });
