+++ packages/client/src/client-v2/index.ts
@@ -1,3 +1,4 @@
 export * from "./api/getAssets";
 export * from "./api/getChains";
 export * from "./setClientOptions";
+export * from "./api/postRoute";
